web_app/requirements.txt
# Web App Dependencies
gradio>=4.0.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
loguru>=0.6.0
